test(sql-pg): move SNI coverage into in-process suite
